refactor: refactor folder structure of static FAQs
